requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
cachetools==5.3.0
charset-normalizer==3.0.1
cloudpickle==1.6.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
gtimer==1.0.0b5
gym==0.19.0
gym-notices==0.0.8
hydra-core==1.3.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
jedi==0.18.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
Pillow==9.4.0
protobuf==4.22.0
pudb==2022.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.1.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
torch==1.9.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
urwid==2.1.2
urwid-readline==0.13
Werkzeug==2.2.3
ydiff==1.2
zipp==3.13.0
